requirements.txt
# Requisitos Básicos (SIAA-2026)
Telethon
moviepy
edge-tts
python-dotenv
streamlit
selenium
requests
pillow
bs4
sqlalchemy
pytest
schedule
fastapi
uvicorn
